Shift filter_outputs, dataset_landing and shared functionality to different files
