+++ mapper/census.go
@@ -9,14 +9,12 @@ import (
 	"net/url"
 	"regexp"
 	"sort"
-	"strconv"
 	"strings"
 
 	"github.com/ONSdigital/dp-api-clients-go/v2/dataset"
 	"github.com/ONSdigital/dp-api-clients-go/v2/filter"
 	"github.com/ONSdigital/dp-api-clients-go/v2/zebedee"
 	"github.com/ONSdigital/dp-frontend-dataset-controller/helpers"
-	"github.com/ONSdigital/dp-frontend-dataset-controller/model"
 	sharedModel "github.com/ONSdigital/dp-frontend-dataset-controller/model"
 	"github.com/ONSdigital/dp-frontend-dataset-controller/model/datasetLandingPageCensus"
 	"github.com/ONSdigital/dp-renderer/helper"
@@ -25,383 +23,19 @@ import (
 
 // Constants...
 const (
-	AlertType           = "alert"
-	CorrectionAlertType = "correction"
-	queryStrKey         = "showAll"
-	Coverage            = "Coverage"
-	FilterOutput        = "_filter_output"
-	AreaType            = "Area type"
-	AnalyticsMaxItems   = 4
+	queryStrKey       = "showAll"
+	Coverage          = "Coverage"
+	AreaType          = "Area type"
+	AnalyticsMaxItems = 4
 )
 
 // CreateCensusDatasetLandingPage creates a census-landing page based on api model responses
 func CreateCensusDatasetLandingPage(isEnableMultivariate bool, ctx context.Context, req *http.Request, basePage coreModel.Page, d dataset.DatasetDetails, version dataset.Version, opts []dataset.Options, initialVersionReleaseDate string, hasOtherVersions bool, allVersions []dataset.Version, latestVersionNumber int, latestVersionURL, lang string, queryStrValues []string, maxNumberOfOptions int, isValidationError, isFilterOutput, hasNoAreaOptions bool, filterOutput map[string]filter.Download, fDims []sharedModel.FilterDimension, serviceMessage string, emergencyBannerContent zebedee.EmergencyBanner) datasetLandingPageCensus.Page {
-	p := datasetLandingPageCensus.Page{
-		Page: basePage,
-	}
-
-	MapCookiePreferences(req, &p.Page.CookiesPreferencesSet, &p.Page.CookiesPolicy)
-
-	// PAGE META-DATA
-	p.Type = d.Type
-	p.Metadata.Title = d.Title
-	p.Language = lang
-	p.URI = req.URL.Path
-	p.Metadata.Description = d.Description
-	p.IsNationalStatistic = d.NationalStatistic
-	p.DatasetId = d.ID
-
-	if d.Contacts != nil && len(*d.Contacts) > 0 {
-		contacts := *d.Contacts
-		if contacts[0].Telephone != "" {
-			p.ContactDetails.Telephone = contacts[0].Telephone
-			p.HasContactDetails = true
-		}
-		if contacts[0].Email != "" {
-			p.ContactDetails.Email = contacts[0].Email
-			p.HasContactDetails = true
-		}
-	}
-
-	p.Version.ReleaseDate = version.ReleaseDate
-	if initialVersionReleaseDate == "" {
-		p.ReleaseDate = p.Version.ReleaseDate
-	} else {
-		p.ReleaseDate = initialVersionReleaseDate
-	}
-
-	p.DatasetLandingPage.Description = strings.Split(d.Description, "\n")
-	p.DatasetLandingPage.HasOtherVersions = hasOtherVersions
-
-	// SITE-WIDE BANNERS
-	p.BetaBannerEnabled = true
-	p.ServiceMessage = serviceMessage
-	p.EmergencyBanner = mapEmergencyBanner(emergencyBannerContent)
-
-	// CENSUS BRANDING
-	p.ShowCensusBranding = d.Survey == "census"
-
-	// BREADCRUMBS
-	p.Breadcrumb = []coreModel.TaxonomyNode{
-		{
-			Title: "Home",
-			URI:   "/",
-		},
-		{
-			Title: "Census",
-			URI:   "/census",
-		},
-	}
-
-	// BACK LINK
-	p.BackTo = coreModel.BackTo{
-		Text: coreModel.Localisation{
-			LocaleKey: "BackToContents",
-			Plural:    4,
-		},
-		AnchorFragment: "toc",
-	}
-
-	// ALERTS
-	if version.Alerts != nil {
-		for _, alert := range *version.Alerts {
-			switch alert.Type {
-			case CorrectionAlertType:
-				p.DatasetLandingPage.Panels = append(p.DatasetLandingPage.Panels, datasetLandingPageCensus.Panel{
-					DisplayIcon: true,
-					Body:        helper.Localise("HasCorrectionNotice", lang, 1),
-					CssClasses:  []string{"ons-u-mt-m", "ons-u-mb-l"},
-				})
-			case AlertType:
-				p.DatasetLandingPage.Panels = append(p.DatasetLandingPage.Panels, datasetLandingPageCensus.Panel{
-					DisplayIcon: true,
-					Body:        helper.Localise("HasAlert", lang, 1, alert.Description),
-					CssClasses:  []string{"ons-u-mt-m", "ons-u-mb-l"},
-				})
-			}
-		}
-	}
-
-	// TABLE OF CONTENTS
-	sections := make(map[string]coreModel.ContentSection)
-	displayOrder := make([]string, 0)
-
-	p.TableOfContents = coreModel.TableOfContents{
-		AriaLabel: coreModel.Localisation{
-			LocaleKey: "ContentsAria",
-			Plural:    1,
-		},
-		Title: coreModel.Localisation{
-			LocaleKey: "Contents",
-			Plural:    1,
-		},
-	}
-
-	sections["summary"] = coreModel.ContentSection{
-		Title: coreModel.Localisation{
-			LocaleKey: "Summary",
-			Plural:    1,
-		},
-	}
-	displayOrder = append(displayOrder, "summary")
-
-	sections["variables"] = coreModel.ContentSection{
-		Title: coreModel.Localisation{
-			LocaleKey: "Variables",
-			Plural:    4,
-		},
-	}
-	displayOrder = append(displayOrder, "variables")
-
-	sections["get-data"] = coreModel.ContentSection{
-		Title: coreModel.Localisation{
-			LocaleKey: "GetData",
-			Plural:    1,
-		},
-	}
-	displayOrder = append(displayOrder, "get-data")
-
-	if p.HasContactDetails {
-		sections["contact"] = coreModel.ContentSection{
-			Title: coreModel.Localisation{
-				LocaleKey: "ContactUs",
-				Plural:    1,
-			},
-		}
-		displayOrder = append(displayOrder, "contact")
-	}
-
-	sections["protecting-personal-data"] = coreModel.ContentSection{
-		Title: coreModel.Localisation{
-			LocaleKey: "ProtectingPersonalDataTitle",
-			Plural:    1,
-		},
-	}
-	displayOrder = append(displayOrder, "protecting-personal-data")
-
-	if hasOtherVersions {
-		sections["version-history"] = coreModel.ContentSection{
-			Title: coreModel.Localisation{
-				LocaleKey: "VersionHistory",
-				Plural:    1,
-			},
-		}
-		displayOrder = append(displayOrder, "version-history")
-	}
-
-	if d.RelatedContent != nil {
-		sections["related-content"] = coreModel.ContentSection{
-			Title: coreModel.Localisation{
-				LocaleKey: "RelatedContentTitle",
-				Plural:    1,
-			},
-		}
-		displayOrder = append(displayOrder, "related-content")
-	}
-
-	p.TableOfContents.Sections = sections
-	p.TableOfContents.DisplayOrder = displayOrder
-
-	// VERSIONS TABLE
-	if hasOtherVersions {
-		for _, ver := range allVersions {
-			var version sharedModel.Version
-			version.VersionNumber = ver.Version
-			version.ReleaseDate = ver.ReleaseDate
-			versionUrl := helpers.DatasetVersionUrl(ver.Links.Dataset.ID, ver.Edition, strconv.Itoa(ver.Version))
-			version.VersionURL = versionUrl
-			version.IsCurrentPage = versionUrl == req.URL.Path
-			mapCorrectionAlert(&ver, &version)
-
-			p.Versions = append(p.Versions, version)
-		}
-
-		sort.Slice(p.Versions, func(i, j int) bool { return p.Versions[i].VersionNumber > p.Versions[j].VersionNumber })
-
-		p.DatasetLandingPage.LatestVersionURL = latestVersionURL
-	}
-
-	// LATEST VERSIONS PANEL
-	if latestVersionNumber != version.Version && hasOtherVersions {
-		p.DatasetLandingPage.Panels = append(p.DatasetLandingPage.Panels, datasetLandingPageCensus.Panel{
-			DisplayIcon: true,
-			Body:        helper.Localise("HasNewVersion", lang, 1, latestVersionURL),
-			CssClasses:  []string{"ons-u-mt-m", "ons-u-mb-l"},
-		})
-	}
-
-	// SHARING LINKS
-	currentUrl := helpers.GetCurrentUrl(lang, p.SiteDomain, req.URL.Path)
-	p.DatasetLandingPage.DatasetURL = currentUrl
-	p.DatasetLandingPage.ShareDetails.Language = lang
-	p.DatasetLandingPage.ShareDetails.ShareLocations = []datasetLandingPageCensus.Share{
-		{
-			Title: "Facebook",
-			Link:  helpers.GenerateSharingLink("facebook", currentUrl, d.Title),
-			Icon:  "facebook",
-		},
-		{
-			Title: "Twitter",
-			Link:  helpers.GenerateSharingLink("twitter", currentUrl, d.Title),
-			Icon:  "twitter",
-		},
-		{
-			Title: "LinkedIn",
-			Link:  helpers.GenerateSharingLink("linkedin", currentUrl, d.Title),
-			Icon:  "linkedin",
-		},
-		{
-			Title: "Email",
-			Link:  helpers.GenerateSharingLink("email", currentUrl, d.Title),
-			Icon:  "email",
-		},
-	}
-
-	// RELATED CONTENT
-	p.DatasetLandingPage.RelatedContentItems = []datasetLandingPageCensus.RelatedContentItem{}
-	if d.RelatedContent != nil {
-		for _, content := range *d.RelatedContent {
-			p.DatasetLandingPage.RelatedContentItems = append(p.DatasetLandingPage.RelatedContentItems, datasetLandingPageCensus.RelatedContentItem{
-				Title: content.Title,
-				Link:  content.HRef,
-				Text:  content.Description,
-			})
-		}
-	}
-
-	// ERRORS
-	if isValidationError {
-		p.Error = coreModel.Error{
-			Title: p.Metadata.Title,
-			ErrorItems: []coreModel.ErrorItem{
-				{
-					Description: coreModel.Localisation{
-						LocaleKey: "GetDataValidationError",
-						Plural:    1,
-					},
-					URL: "#select-format-error",
-				},
-			},
-			Language: lang,
-		}
-	}
-
-	/*
-		--------------------------------
-		FILTER FLEX CONTENT BELOW THIS LINE
-		--------------------------------
-	*/
-
-	isFlex := strings.Contains(d.Type, "flex")
-	isMultivariate := strings.Contains(d.Type, "multivariate") && isEnableMultivariate
-	p.DatasetLandingPage.IsMultivariate = isMultivariate
-	p.DatasetLandingPage.IsFlexibleForm = isFlex || isMultivariate
-
-	/*
-		--------------------------------
-		isFilterOutput DEPENDENT CONTENT BELOW THIS LINE
-		--------------------------------
-	*/
-
 	if isFilterOutput {
-		p.Type += FilterOutput
-		p.SearchNoIndexEnabled = true
-
-		// DOWNLOADS
-		for ext, download := range filterOutput {
-			p.Version.Downloads = append(p.Version.Downloads, sharedModel.Download{
-				Extension: strings.ToLower(ext),
-				Size:      download.Size,
-				URI:       download.URL,
-			})
-		}
-		p.Version.Downloads = orderDownloads(p.Version.Downloads)
-
-		if len(filterOutput) >= 3 {
-			p.DatasetLandingPage.HasDownloads = true
-			p.DatasetLandingPage.ShowXLSXInfo = true
-		}
-
-		// DIMENSIONS
-		p.DatasetLandingPage.Dimensions = mapFilterOutputDims(fDims, queryStrValues, req.URL.Path, isMultivariate)
-		coverage := []sharedModel.Dimension{
-			{
-				IsCoverage:        true,
-				IsDefaultCoverage: hasNoAreaOptions,
-				Title:             Coverage,
-				Name:              strings.ToLower(Coverage),
-				ID:                strings.ToLower(Coverage),
-				Values:            fDims[0].Options,
-				ShowChange:        true,
-			},
-		}
-		temp := append(coverage, p.DatasetLandingPage.Dimensions[1:]...)
-		p.DatasetLandingPage.Dimensions = append(p.DatasetLandingPage.Dimensions[:1], temp...)
-
-		// COLLAPSIBLE CONTENT
-		p.Collapsible = coreModel.Collapsible{
-			Title: coreModel.Localisation{
-				LocaleKey: "VariablesExplanation",
-				Plural:    4,
-			},
-			CollapsibleItems: populateCollapsible(version.Dimensions, true),
-		}
-
-		// ANALYTICS
-		p.PreGTMJavaScript = append(p.PreGTMJavaScript, getDataLayerJavaScript(getFilterAnalytics(fDims, hasNoAreaOptions)))
+		return CreateCensusFilterOutputsPage(isEnableMultivariate, ctx, req, basePage, d, version, opts, initialVersionReleaseDate, hasOtherVersions, allVersions, latestVersionNumber, latestVersionURL, lang, queryStrValues, maxNumberOfOptions, isValidationError, isFilterOutput, hasNoAreaOptions, filterOutput, fDims, serviceMessage, emergencyBannerContent)
 	} else {
-
-		// DOWNLOADS
-		for ext, download := range version.Downloads {
-			p.Version.Downloads = append(p.Version.Downloads, sharedModel.Download{
-				Extension: strings.ToLower(ext),
-				Size:      download.Size,
-				URI:       download.URL,
-			})
-		}
-		p.Version.Downloads = orderDownloads(p.Version.Downloads)
-
-		if len(version.Downloads) >= 3 {
-			p.DatasetLandingPage.HasDownloads = true
-		}
-
-		// DIMENSIONS
-		if len(opts) > 0 {
-			p.DatasetLandingPage.Dimensions, p.DatasetLandingPage.QualityStatements = mapCensusOptionsToDimensions(version.Dimensions, opts, queryStrValues, req.URL.Path, lang, isFlex, isMultivariate)
-			coverage := []sharedModel.Dimension{
-				{
-					IsCoverage:        true,
-					IsDefaultCoverage: true,
-					Title:             Coverage,
-					Name:              strings.ToLower(Coverage),
-					ShowChange:        isFlex || isMultivariate,
-					ID:                strings.ToLower(Coverage),
-				},
-			}
-			temp := append(coverage, p.DatasetLandingPage.Dimensions[1:]...)
-			p.DatasetLandingPage.Dimensions = append(p.DatasetLandingPage.Dimensions[:1], temp...)
-		}
-
-		// COLLAPSIBLE
-		p.Collapsible = coreModel.Collapsible{
-			Title: coreModel.Localisation{
-				LocaleKey: "VariablesExplanation",
-				Plural:    4,
-			},
-			CollapsibleItems: populateCollapsible(version.Dimensions, false),
-		}
-
-		// ANALYTICS
-		p.PreGTMJavaScript = append(p.PreGTMJavaScript, getDataLayerJavaScript(getAnalytics(p.DatasetLandingPage.Dimensions)))
-
-		// FINAL FORMATTING
-		if len(p.DatasetLandingPage.QualityStatements) > 0 {
-			qsLen := len(p.DatasetLandingPage.QualityStatements)
-			p.DatasetLandingPage.QualityStatements[qsLen-1].CssClasses = append(p.DatasetLandingPage.QualityStatements[qsLen-1].CssClasses, "ons-u-mb-l")
-		}
+		return CreateCensusLandingPage(isEnableMultivariate, ctx, req, basePage, d, version, opts, initialVersionReleaseDate, hasOtherVersions, allVersions, latestVersionNumber, latestVersionURL, lang, queryStrValues, maxNumberOfOptions, isValidationError, isFilterOutput, hasNoAreaOptions, filterOutput, fDims, serviceMessage, emergencyBannerContent)
 	}
-
-	return p
 }
 
 func orderDownloads(downloads []sharedModel.Download) []sharedModel.Download {
@@ -593,53 +227,4 @@ func cleanDimensionLabel(label string) string {
 func getDataLayerJavaScript(analytics map[string]string) template.JS {
 	jsonStr, _ := json.Marshal(analytics)
 	return template.JS(`dataLayer.push(` + string(jsonStr) + `);`)
-}
-
-func getAnalytics(dimensions []model.Dimension) map[string]string {
-	analytics := make(map[string]string, 5)
-	var dimensionIDs []string
-	for _, dimension := range dimensions {
-		if dimension.IsAreaType {
-			analytics["areaType"] = dimension.ID
-			analytics["coverageCount"] = "0"
-		} else if !dimension.IsCoverage {
-			dimensionIDs = append(dimensionIDs, dimension.ID)
-		}
-	}
-	analytics["dimensions"] = strings.Join(dimensionIDs, ",")
-
-	return analytics
-}
-
-func getFilterAnalytics(filterDimensions []sharedModel.FilterDimension, defaultCoverage bool) map[string]string {
-	analytics := make(map[string]string, 5)
-	var dimensionIDs []string
-	for _, filterDimension := range filterDimensions {
-		dimension := filterDimension.ModelDimension
-		if dimension.IsAreaType != nil && *dimension.IsAreaType {
-			analytics["areaType"] = dimension.ID
-
-			if defaultCoverage {
-				analytics["coverageCount"] = "0"
-			} else {
-				analytics["coverageCount"] = strconv.Itoa(len(dimension.Options))
-
-				if len(dimension.Options) > 0 {
-					if len(dimension.Options) <= AnalyticsMaxItems {
-						analytics["coverage"] = strings.Join(dimension.Options, ",")
-					}
-					if dimension.FilterByParent == "" {
-						analytics["coverageAreaType"] = dimension.ID
-					} else {
-						analytics["coverageAreaType"] = dimension.FilterByParent
-					}
-				}
-			}
-		} else {
-			dimensionIDs = append(dimensionIDs, dimension.ID)
-		}
-	}
-	analytics["dimensions"] = strings.Join(dimensionIDs, ",")
-
-	return analytics
 }
